requirements.txt
psycopg2==2.9.6
python-decouple==3.8
Django==3.2.10
djangorestframework==3.12.4
requests==2.28.2
django-environ==0.8.1
python-dotenv==0.19.2
